util script and EdgeTable class
