+++ setup.py
@@ -8,7 +8,7 @@ with open("requirement.txt", 'r') as req:
 
 setuptools.setup(
     name="edge_pydb",
-    version="0.1.15",
+    version="0.1.18",
     author="Tony Wong",
     author_email="tonywong94@gmail.com",
     description="Python based database for CARMA EDGE",
